requirements.txt
langchain-core>=0.1.0
langgraph>=0.0.20
langchain-anthropic>=0.1.1
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain-mcp-adapters>=0.0.1 